Feat : 비밀번호 검증 기능(수정,삭제) 추가
비밀번호를 사용하여 수정 및 삭제 검증 기능 추가
(비밀번호가 틀리면 데이터가 수정, 삭제가 되지 않음.)

+++ src/main/java/com/example/schedulemanagement/service/ScheduleServiceImpl.java
@@ -7,9 +7,11 @@ import com.example.schedulemanagement.repository.ScheduleRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -43,33 +45,36 @@ public class ScheduleServiceImpl implements ScheduleService {
 
     @Transactional
     @Override
-    public ScheduleResponseDto updateSchedule(Long id, String userName, String title, String contents) {
-        if (title == null || userName == null) {
+    public ScheduleResponseDto updateSchedule(Long id, ScheduleRequestDto dto) {
+        if (dto.getTitle() == null || dto.getUserName() == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Username and title cannot be null");
         }
-//        if (!schedule.getPassword().equals(password)) {
-//            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Passwords don't match");
-//        }
-        int updatedRow = scheduleRepository.updateSchedule(id, userName, title, contents);
+        Schedule schedule = scheduleRepository.findScheduleByIdOrElseThrow(id);
+
+        if (!schedule.getPassword().equals(dto.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Passwords don't match");
+        }
+
+        int updatedRow = scheduleRepository.updateSchedule(id, dto.getTitle(), dto.getContents(), dto.getUserName());
 
         if (updatedRow == 0) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Schedule not found");
         }
-        Schedule schedule = scheduleRepository.findScheduleByIdOrElseThrow(id);
+        Schedule updatedSchedule = scheduleRepository.findScheduleByIdOrElseThrow(id);
 
-        return new ScheduleResponseDto(schedule);
+        return new ScheduleResponseDto(updatedSchedule);
     }
 
     @Override
-    public void deleteSchedule(Long id, ScheduleRequestDto dto) {
+    public void deleteSchedule(Long id, @RequestBody ScheduleRequestDto dto) {
+        Schedule schedule = scheduleRepository.findScheduleByIdOrElseThrow(id);
+        if (!schedule.getPassword().equals(dto.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Passwords don't match");
+        }
         int deleteRow = scheduleRepository.deleteSchedule(id);
 
         if (deleteRow == 0) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Schedule not found");
         }
-
-//        if (!schedule.getPassword().equals(dto.getPassword())) {
-//            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Passwords don't match");
-//        }
     }
 }
